Working Version
Created a new branch for this - got too deep in Git Hell.

It works but not efficient enough. Need to tidy up how some of it works
since for large event files too much is saved in internal memory and
it fills up and crashes. Testing was done on an event file from 1987 -
1996 and it crashed in 1991.